backend/requirements.txt
flask
python-dotenv
azure-identity
azure-mgmt-containerinstance